Add improved isolation
The extension content is only injected into classify page when the user
navigates to it directly. Since Zooniverse is implemented as a
single-page app, clicking an internal link to another page does not
unload the extension content.

Add namespacing to css to restrict application to the Rainfall Rescue
classify page.

Feature initialization functions now return a callback that, when
called, removes the feature from the page.

Use mutation observer to watch for changes in the url, and add/remove
extension features as necessary.